task_final/scripts/requirements.txt
pymongo
faker
tqdmv
matplotlib 
numpy 